Move constructor to resource package
