+++ internal/reconstitution/api.go
@@ -2,17 +2,15 @@ package reconstitution
 
 import (
 	"context"
-	"sync"
-	"time"
 
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 
 	apiv1 "github.com/Azure/eno/api/v1"
-	"github.com/Azure/eno/internal/readiness"
+	"github.com/Azure/eno/internal/resource"
 )
+
+type Resource = resource.Resource
 
 // Reconciler is implemented by types that can reconcile individual, reconstituted resources.
 type Reconciler interface {
@@ -22,7 +20,7 @@ type Reconciler interface {
 
 // Client provides read/write access to a collection of reconstituted resources.
 type Client interface {
-	Get(ctx context.Context, comp *CompositionRef, res *ResourceRef) (*Resource, bool)
+	Get(ctx context.Context, comp *CompositionRef, res *resource.Ref) (*resource.Resource, bool)
 	PatchStatusAsync(ctx context.Context, req *ManifestRef, patchFn StatusPatchFn)
 }
 
@@ -42,30 +40,6 @@ func (m *ManifestRef) FindStatus(slice *apiv1.ResourceSlice) *apiv1.ResourceStat
 	return &state
 }
 
-// Resource is the controller's internal representation of a single resource out of a ResourceSlice.
-type Resource struct {
-	lastSeenMeta
-	lastReconciledMeta
-
-	Ref             ResourceRef
-	Manifest        *apiv1.Manifest
-	GVK             schema.GroupVersionKind
-	SliceDeleted    bool
-	ReadinessChecks readiness.Checks
-}
-
-func (r *Resource) Deleted() bool { return r.SliceDeleted || r.Manifest.Deleted }
-
-func (r *Resource) Parse() (*unstructured.Unstructured, error) {
-	u := &unstructured.Unstructured{}
-	return u, u.UnmarshalJSON([]byte(r.Manifest.Manifest))
-}
-
-// ResourceRef refers to a specific synthesized resource.
-type ResourceRef struct {
-	Name, Namespace, Kind string
-}
-
 // CompositionRef refers to a specific generation of a composition.
 type CompositionRef struct {
 	Name, Namespace string
@@ -83,50 +57,7 @@ func NewCompositionRef(comp *apiv1.Composition) *CompositionRef {
 // Request is like controller-runtime reconcile.Request but for reconstituted resources.
 // https://pkg.go.dev/sigs.k8s.io/controller-runtime/pkg/reconcile#Request
 type Request struct {
-	Resource    ResourceRef
+	Resource    resource.Ref
 	Manifest    ManifestRef
 	Composition types.NamespacedName
-}
-
-type lastSeenMeta struct {
-	lock            sync.Mutex
-	resourceVersion string
-}
-
-func (l *lastSeenMeta) ObserveVersion(rv string) {
-	l.lock.Lock()
-	defer l.lock.Unlock()
-	l.resourceVersion = rv
-}
-
-func (l *lastSeenMeta) HasBeenSeen() bool {
-	l.lock.Lock()
-	defer l.lock.Unlock()
-	return l.resourceVersion != ""
-}
-
-func (l *lastSeenMeta) MatchesLastSeen(rv string) bool {
-	l.lock.Lock()
-	defer l.lock.Unlock()
-	return l.resourceVersion == rv
-}
-
-type lastReconciledMeta struct {
-	lock           sync.Mutex
-	lastReconciled *time.Time
-}
-
-func (l *lastReconciledMeta) ObserveReconciliation() time.Duration {
-	now := time.Now()
-
-	l.lock.Lock()
-	defer l.lock.Unlock()
-
-	var latency time.Duration
-	if l.lastReconciled != nil {
-		latency = now.Sub(*l.lastReconciled)
-	}
-
-	l.lastReconciled = &now
-	return latency
 }
